refactor : 이벤트 페이로드 id Long-> String으로 수정

+++ service/memo-service/src/main/kotlin/jeonghyeon/msa/memo/kafka/consumer/UsersCreateEventConsumer.kt
@@ -29,7 +29,7 @@ class UsersCreateEventConsumer(
         if (event != null) {
             if (event.type == EventType.AUTH_CREATE) {
                 val payload: AuthCreateEventPayload = event.payload as AuthCreateEventPayload
-                memoService.createUser(UsersDto(payload.usersId, payload.username))
+                memoService.createUser(UsersDto(payload.usersId.toLong(), payload.username))
                 ack.acknowledge()
                 return
             }
